+++ packages/core/CHANGELOG.md
@@ -3,9 +3,132 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [2.1.1](https://github.com/airgram/airgram/compare/@airgram/core@2.1.0...@airgram/core@2.1.1) (2019-11-07)
+# [2.3.0-next.8](https://github.com/esindger/airgram/compare/@airgram/core@2.3.0-next.7...@airgram/core@2.3.0-next.8) (2019-11-12)
+
+
+### Bug Fixes
+
+* **core:** context state ([9086e41](https://github.com/esindger/airgram/commit/9086e41))
+
+
+
+
+
+# [2.3.0-next.7](https://github.com/esindger/airgram/compare/@airgram/core@2.3.0-next.6...@airgram/core@2.3.0-next.7) (2019-11-12)
+
+
+### Bug Fixes
+
+* **core:** context state ([6afc22b](https://github.com/esindger/airgram/commit/6afc22b))
+
+
+### Features
+
+* bump ([7a64648](https://github.com/esindger/airgram/commit/7a64648))
+* bump ([cc5ffd4](https://github.com/esindger/airgram/commit/cc5ffd4))
+
+
+
+
+
+# [2.3.0-next.6](https://github.com/esindger/airgram/compare/@airgram/core@2.3.0-next.5...@airgram/core@2.3.0-next.6) (2019-11-12)
+
+
+### Features
+
+* bump ([b869fc8](https://github.com/esindger/airgram/commit/b869fc8))
+
+
+
+
+
+# [2.3.0-next.5](https://github.com/esindger/airgram/compare/@airgram/core@2.3.0-next.4...@airgram/core@2.3.0-next.5) (2019-10-20)
 
 **Note:** Version bump only for package @airgram/core
+
+
+
+
+
+# [2.3.0-next.4](https://github.com/esindger/airgram/compare/@airgram/core@2.3.0-next.3...@airgram/core@2.3.0-next.4) (2019-10-20)
+
+
+### Features
+
+* **core:** add destroy method ([f83afe3](https://github.com/esindger/airgram/commit/f83afe3))
+
+
+
+
+
+# [2.3.0-next.3](https://github.com/esindger/airgram/compare/@airgram/core@2.3.0-next.2...@airgram/core@2.3.0-next.3) (2019-10-19)
+
+**Note:** Version bump only for package @airgram/core
+
+
+
+
+
+# [2.3.0-next.2](https://github.com/esindger/airgram/compare/@airgram/core@2.3.0-next.1...@airgram/core@2.3.0-next.2) (2019-10-19)
+
+
+### Bug Fixes
+
+* **vue:** tsconfig dom lib support ([dd3b61e](https://github.com/esindger/airgram/commit/dd3b61e))
+
+
+
+
+
+# [2.3.0-next.1](https://github.com/esindger/airgram/compare/@airgram/core@2.3.0-next.0...@airgram/core@2.3.0-next.1) (2019-10-19)
+
+
+### Features
+
+* **web:** web worker support ([8266f30](https://github.com/esindger/airgram/commit/8266f30))
+
+
+
+
+
+# [2.3.0-next.0](https://github.com/esindger/airgram/compare/@airgram/core@2.2.1...@airgram/core@2.3.0-next.0) (2019-10-14)
+
+
+### Features
+
+* **core:** add promise property to the API response object ([1f5c981](https://github.com/esindger/airgram/commit/1f5c981))
+
+
+
+
+
+## [2.2.1](https://github.com/esindger/airgram/compare/@airgram/core@2.2.0...@airgram/core@2.2.1) (2019-09-29)
+
+**Note:** Version bump only for package @airgram/core
+
+
+
+
+
+# 2.2.0 (2019-09-29)
+
+
+### Bug Fixes
+
+* do not freeze updates while initialization parameters are sending ([b2cc07b](https://github.com/esindger/airgram/commit/b2cc07b))
+* serialize `api` ([1d8f6a4](https://github.com/esindger/airgram/commit/1d8f6a4))
+
+
+### Features
+
+* bump to v2, bump to TDLib 1.5 ([4cb51bb](https://github.com/esindger/airgram/commit/4cb51bb))
+* improve middleware typings, add `on` helper ([be5c406](https://github.com/esindger/airgram/commit/be5c406))
+* improve typings ([31d7d3f](https://github.com/esindger/airgram/commit/31d7d3f))
+
+
+### BREAKING CHANGES
+
+* see upgrade guide