+++ packages/use-models/CHANGELOG.md
@@ -3,9 +3,120 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [2.0.3](https://github.com/airgram/airgram/compare/@airgram/use-models@2.0.2...@airgram/use-models@2.0.3) (2019-11-07)
+# [2.2.0-next.1](https://github.com/esindger/airgram/compare/@airgram/use-models@2.2.0-next.0...@airgram/use-models@2.2.0-next.1) (2019-11-12)
 
 **Note:** Version bump only for package @airgram/use-models
+
+
+
+
+
+# [2.2.0-next.0](https://github.com/esindger/airgram/compare/@airgram/use-models@2.1.3-next.6...@airgram/use-models@2.2.0-next.0) (2019-11-12)
+
+
+### Features
+
+* bump ([7a64648](https://github.com/esindger/airgram/commit/7a64648))
+* bump ([cc5ffd4](https://github.com/esindger/airgram/commit/cc5ffd4))
+
+
+
+
+
+## [2.1.3-next.6](https://github.com/esindger/airgram/compare/@airgram/use-models@2.1.3-next.5...@airgram/use-models@2.1.3-next.6) (2019-11-12)
+
+**Note:** Version bump only for package @airgram/use-models
+
+
+
+
+
+## [2.1.3-next.5](https://github.com/esindger/airgram/compare/@airgram/use-models@2.1.3-next.4...@airgram/use-models@2.1.3-next.5) (2019-10-20)
+
+**Note:** Version bump only for package @airgram/use-models
+
+
+
+
+
+## [2.1.3-next.4](https://github.com/esindger/airgram/compare/@airgram/use-models@2.1.3-next.3...@airgram/use-models@2.1.3-next.4) (2019-10-20)
+
+**Note:** Version bump only for package @airgram/use-models
+
+
+
+
+
+## [2.1.3-next.3](https://github.com/esindger/airgram/compare/@airgram/use-models@2.1.3-next.2...@airgram/use-models@2.1.3-next.3) (2019-10-19)
+
+**Note:** Version bump only for package @airgram/use-models
+
+
+
+
+
+## [2.1.3-next.2](https://github.com/esindger/airgram/compare/@airgram/use-models@2.1.3-next.1...@airgram/use-models@2.1.3-next.2) (2019-10-19)
+
+
+### Bug Fixes
+
+* **vue:** tsconfig dom lib support ([dd3b61e](https://github.com/esindger/airgram/commit/dd3b61e))
+
+
+
+
+
+## [2.1.3-next.1](https://github.com/esindger/airgram/compare/@airgram/use-models@2.1.3-next.0...@airgram/use-models@2.1.3-next.1) (2019-10-19)
+
+**Note:** Version bump only for package @airgram/use-models
+
+
+
+
+
+## [2.1.3-next.0](https://github.com/esindger/airgram/compare/@airgram/use-models@2.1.2...@airgram/use-models@2.1.3-next.0) (2019-10-14)
+
+**Note:** Version bump only for package @airgram/use-models
+
+
+
+
+
+## [2.1.2](https://github.com/esindger/airgram/compare/@airgram/use-models@2.1.2-next.0...@airgram/use-models@2.1.2) (2019-09-29)
+
+**Note:** Version bump only for package @airgram/use-models
+
+
+
+
+
+## [2.1.2-next.0](https://github.com/esindger/airgram/compare/@airgram/use-models@2.1.1...@airgram/use-models@2.1.2-next.0) (2019-09-29)
+
+**Note:** Version bump only for package @airgram/use-models
+
+
+
+
+
+## [2.1.1](https://github.com/esindger/airgram/compare/@airgram/use-models@2.1.0...@airgram/use-models@2.1.1) (2019-09-29)
+
+**Note:** Version bump only for package @airgram/use-models
+
+
+
+
+
+# 2.1.0 (2019-09-29)
+
+
+### Features
+
+* bump to v2, bump to TDLib 1.5 ([4cb51bb](https://github.com/esindger/airgram/commit/4cb51bb))
+
+
+### BREAKING CHANGES
+
+* see upgrade guide