+++ packages/web/CHANGELOG.md
@@ -3,9 +3,142 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [2.1.1](https://github.com/airgram/airgram/compare/@airgram/web@2.1.0...@airgram/web@2.1.1) (2019-11-07)
+# [2.3.0-next.7](https://github.com/esindger/airgram/compare/@airgram/web@2.3.0-next.6...@airgram/web@2.3.0-next.7) (2019-11-12)
 
 **Note:** Version bump only for package @airgram/web
+
+
+
+
+
+# [2.3.0-next.6](https://github.com/esindger/airgram/compare/@airgram/web@2.3.0-next.5...@airgram/web@2.3.0-next.6) (2019-11-12)
+
+
+### Features
+
+* bump ([7a64648](https://github.com/esindger/airgram/commit/7a64648))
+* bump ([cc5ffd4](https://github.com/esindger/airgram/commit/cc5ffd4))
+
+
+
+
+
+# [2.3.0-next.5](https://github.com/esindger/airgram/compare/@airgram/web@2.3.0-next.4...@airgram/web@2.3.0-next.5) (2019-11-12)
+
+
+### Features
+
+* bump ([b869fc8](https://github.com/esindger/airgram/commit/b869fc8))
+
+
+
+
+
+# [2.3.0-next.4](https://github.com/esindger/airgram/compare/@airgram/web@2.3.0-next.3...@airgram/web@2.3.0-next.4) (2019-10-20)
+
+**Note:** Version bump only for package @airgram/web
+
+
+
+
+
+# [2.3.0-next.3](https://github.com/esindger/airgram/compare/@airgram/web@2.3.0-next.2...@airgram/web@2.3.0-next.3) (2019-10-20)
+
+
+### Bug Fixes
+
+* **web:** drop support models due models can not be transferred from web worker ([98cfcc9](https://github.com/esindger/airgram/commit/98cfcc9))
+
+
+
+
+
+# [2.3.0-next.2](https://github.com/esindger/airgram/compare/@airgram/web@2.3.0-next.1...@airgram/web@2.3.0-next.2) (2019-10-19)
+
+**Note:** Version bump only for package @airgram/web
+
+
+
+
+
+# [2.3.0-next.1](https://github.com/esindger/airgram/compare/@airgram/web@2.3.0-next.0...@airgram/web@2.3.0-next.1) (2019-10-19)
+
+
+### Bug Fixes
+
+* **vue:** tsconfig dom lib support ([dd3b61e](https://github.com/esindger/airgram/commit/dd3b61e))
+
+
+
+
+
+# [2.3.0-next.0](https://github.com/esindger/airgram/compare/@airgram/web@2.2.3-next.1...@airgram/web@2.3.0-next.0) (2019-10-19)
+
+
+### Features
+
+* **web:** web worker support ([8266f30](https://github.com/esindger/airgram/commit/8266f30))
+
+
+
+
+
+## [2.2.3-next.1](https://github.com/esindger/airgram/compare/@airgram/web@2.2.3-next.0...@airgram/web@2.2.3-next.1) (2019-10-14)
+
+**Note:** Version bump only for package @airgram/web
+
+
+
+
+
+## [2.2.3-next.0](https://github.com/esindger/airgram/compare/@airgram/web@2.2.2...@airgram/web@2.2.3-next.0) (2019-10-12)
+
+
+### Bug Fixes
+
+* **web:** snakeCase function ([e3a509b](https://github.com/esindger/airgram/commit/e3a509b))
+
+
+
+
+
+## [2.2.2](https://github.com/esindger/airgram/compare/@airgram/web@2.2.2-next.0...@airgram/web@2.2.2) (2019-09-29)
+
+**Note:** Version bump only for package @airgram/web
+
+
+
+
+
+## [2.2.2-next.0](https://github.com/esindger/airgram/compare/@airgram/web@2.2.1...@airgram/web@2.2.2-next.0) (2019-09-29)
+
+**Note:** Version bump only for package @airgram/web
+
+
+
+
+
+## [2.2.1](https://github.com/esindger/airgram/compare/@airgram/web@2.2.0...@airgram/web@2.2.1) (2019-09-29)
+
+**Note:** Version bump only for package @airgram/web
+
+
+
+
+
+# 2.2.0 (2019-09-29)
+
+
+### Features
+
+* bump to v2, bump to TDLib 1.5 ([4cb51bb](https://github.com/esindger/airgram/commit/4cb51bb))
+* export config interface ([f0e82e0](https://github.com/esindger/airgram/commit/f0e82e0))
+* improve middleware typings, add `on` helper ([be5c406](https://github.com/esindger/airgram/commit/be5c406))
+
+
+### BREAKING CHANGES
+
+* see upgrade guide